Separated serial_port from driver code to be a standalone lib
