+++ hans_cute_driver/CMakeLists.txt
@@ -10,6 +10,8 @@ add_compile_options(-std=c++17)
 find_package(catkin REQUIRED COMPONENTS
 )
 
+add_subdirectory(custom_serial_port)
+
 # # System dependencies are found with CMake's conventions
 # find_package(Boost REQUIRED COMPONENTS system)
 
@@ -100,8 +102,8 @@ find_package(catkin REQUIRED COMPONENTS
 # # DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
 
-  INCLUDE_DIRS include
-  LIBRARIES ${PROJECT_NAME}
+  INCLUDE_DIRS include custom_serial_port/include
+  LIBRARIES ${PROJECT_NAME} custom_serial_port
   CATKIN_DEPENDS roscpp
 
   # DEPENDS system_lib
@@ -110,8 +112,8 @@ catkin_package(
 # ##########
 # # Build ##
 # ##########
+
 set(${PROJECT_NAME}_SOURCES
-  src/custom_serial_port.cpp
   src/serial_command.cpp
   src/hans_cute_serial.cpp
   src/hans_cute_driver.cpp
@@ -123,6 +125,7 @@ set(${PROJECT_NAME}_SOURCES
 include_directories(
   include
   ${catkin_INCLUDE_DIRS}
+  custom_serial_port/include
 )
 
 # # Declare a C++ library
@@ -136,6 +139,7 @@ add_dependencies(${PROJECT_NAME} ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EX
 # # Specify libraries to link a library or executable target against
 target_link_libraries(${PROJECT_NAME}
   ${catkin_LIBRARIES}
+  custom_serial_port
 )
 
 # # Declare a C++ executable
@@ -156,6 +160,7 @@ add_dependencies(${PROJECT_NAME}_node ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catk
 # # Specify libraries to link a library or executable target against
 target_link_libraries(${PROJECT_NAME}_node
   ${catkin_LIBRARIES}
+  custom_serial_port
 )
 
 # ############
@@ -210,7 +215,7 @@ target_link_libraries(${PROJECT_NAME}_node
 # catkin_add_gtest(${PROJECT_NAME}-test ${${PROJECT_NAME}_TESTS})
 
 # if(TARGET ${PROJECT_NAME}-test)
-#   target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME})
+# target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME})
 # endif()
 
 # # Add folders to be run by python nosetests
