+++ hans_cute_driver/custom_serial_port/src/dummy_serial_port.cpp
@@ -1,0 +1,9 @@
+#include "custom_serial_port/dummy_serial_port.h"
+
+DummySerialPort::DummySerialPort(const std::string &port, const unsigned int &baud_rate)
+{
+}
+
+DummySerialPort ~DummySerialPort()
+{
+}
