+++ hans_cute_driver/package.xml
@@ -51,10 +51,13 @@
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>roscpp</build_depend>
   <build_depend>serial</build_depend>
+  <build_depend>custom_serial_port</build_depend>
   <build_export_depend>roscpp</build_export_depend>
   <build_export_depend>serial</build_export_depend>
+  <build_export_depend>custom_serial_port</build_export_depend>
   <exec_depend>roscpp</exec_depend>
   <exec_depend>serial</exec_depend>
+  <exec_depend>custom_serial_port</exec_depend>
 
 
   <!-- The export tag contains other, unspecified, tags -->
